Add ability to test what type of error a node has
We use `assert_current_node` to test at what node we are in the flow:
```
assert_current_node(:how_many_days_per_week?)
```
We sometimes want to assert that there's an error at that point. We can
do it by passing a second argument like so:
```
assert_current_node(:how_many_days_per_week?, error: true)
```
This is fine but I want to test what the error is, not simply that there
is one. With this change, we can now pass an error class as value for
the `error` key, if we want, like so:
```
assert_current_node(:how_many_days_per_week?, error: SmartAnswer::InvalidResponse)
```
This doesn't break previous tests that use a boolean as value for the
`error` key.

Trello card: https://trello.com/c/Me3aeYlL/1363-8-update-the-days-worked-per-week-flow-for-the-holiday-entitlement-calculator

Correct String interpolation