Set leave year start to employment start date for partial years
In the scenario where the leave year has not been set, we should
use the first working day as the beginning of the leave year rather
than the actual beginning of year. This ensures leap years will not
be included in working ranges that do not actually cover February 29th.